requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2 